feat(refactor): 폴더링
