+++ app/src/main/java/org/android/go/sopt/domain/model/Music.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.data.model
+package org.android.go.sopt.domain.model
 
 import androidx.annotation.DrawableRes
